fix test for user controller

+++ user-microservice/src/main/java/nl/tudelft/sem/template/controllers/UserController.java
@@ -1,8 +1,8 @@
 package nl.tudelft.sem.template.controllers;
 
-import java.util.NoSuchElementException;
 import javax.transaction.Transactional;
 import javax.validation.Valid;
+import java.util.NoSuchElementException;
 import nl.tudelft.sem.template.api.UserApi;
 import nl.tudelft.sem.template.authentication.AuthManager;
 import nl.tudelft.sem.template.domain.user.AppUser;
@@ -29,6 +29,7 @@ public class UserController implements UserApi {
      * Instantiates a new User controller.
      *
      * @param userService used to manage user services
+     * @param authManager Used for authentication-related checks.
      */
     @Autowired
     public UserController(UserService userService, AuthManager authManager) {
@@ -38,10 +39,10 @@ public class UserController implements UserApi {
 
     /**
      * Checks whether a user with the given id exists, retrieves it if yes.
-
+     *
      * @param userId - id of the to be found user
      * @return - bad request if invalid id, unauthorized access if expired token,
-     *           not found if user not found, appUser if user found
+     * not found if user not found, appUser if user found
      */
     @Override
     @Transactional
@@ -55,12 +56,13 @@ public class UserController implements UserApi {
         }
         return ResponseEntity.ok(userService.getUserById(userId).toModelUser());
     }
+
     /**
      * Checks whether a user with the given email address exists, retrieves it if yes.
-
+     *
      * @param email - email of the to be found user
      * @return - bad request if invalid email, unauthorized access if expired token,
-     *           not found if user not found, appUser if user found
+     * not found if user not found, appUser if user found
      */
     @Override
     @Transactional
@@ -83,13 +85,14 @@ public class UserController implements UserApi {
      */
     @Override
     @Transactional
-    @PreAuthorize("@RoleService.hasPermission(userService, authManager, attendeeService, "
-            + "#track.getEventId(), #track.getId(), 0)") // 401
     public ResponseEntity<User> createAccount(@Valid @RequestBody User user) {
         try {
-            // if (!userService.userExistsByEmail(new Email(authManager.getEmail()))) {
-            //     return new ResponseEntity<>(HttpStatus.UNAUTHORIZED); // 401
-            // }
+            if (!userService.userExistsByEmail(new Email(authManager.getEmail()))) {
+                return new ResponseEntity<>(HttpStatus.UNAUTHORIZED); // 401
+            }
+            if (user == null) {
+                return new ResponseEntity<>(HttpStatus.BAD_REQUEST); // 400
+            }
             userService.createUser(new AppUser(user));
             return ResponseEntity.ok(user); // 200
         } catch (IllegalArgumentException e) {
@@ -99,14 +102,12 @@ public class UserController implements UserApi {
                 return ResponseEntity.status(409).build(); // 409, user already exists
             }
         }
-
     }
 
     /**
      * This method updates an existing User Account.
      *
      * @param updatedUser - user account to be updated
-     *
      * @return responseEntity of method
      */
     @Override
@@ -133,10 +134,16 @@ public class UserController implements UserApi {
      */
     @Override
     @Transactional
-    @PreAuthorize("@RoleService.hasPermission(userService, authManager, attendeeService, "
-            + "#track.getEventId(), #track.getId(), 0)") // 401
     public ResponseEntity<Void> deleteAccount(@PathVariable("userID") Long userId) {
         try {
+            Email email = new Email(authManager.getEmail());
+            if (!userService.userExistsByEmail(email)) {
+                return new ResponseEntity<>(HttpStatus.UNAUTHORIZED); // 401
+            }
+            AppUser user = userService.getUserByEmail(email);
+            if (user.getId() != userId){
+                return new ResponseEntity<>(HttpStatus.UNAUTHORIZED); // 401
+            }
             userService.deleteUser(userId);
             return new ResponseEntity<>(HttpStatus.NO_CONTENT); // 204
         } catch (IllegalArgumentException e) {
